Move sandboxed runModule to Visual Editor.
