+++ packages/visual-editor/src/utils/kit-loader.ts
@@ -4,14 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { KitConstructor, Kit, asRuntimeKit } from "@google-labs/breadboard";
+import { Kit } from "@google-labs/breadboard";
 import { load } from "@google-labs/breadboard/kits";
 
-export const loadKits = async (kiConstructors: KitConstructor<Kit>[]) => {
-  const loadedKits = kiConstructors.map((kitConstructor) =>
-    asRuntimeKit(kitConstructor)
-  );
-
+export const loadKits = async (loadedKits: Kit[]) => {
   const base = new URL(`${self.location.origin}/kits.json`);
   const response = await fetch(base);
   const kitList = await response.json();
